relayer/requirements.txt
aioimaplib==1.1.0
aiosmtplib==3.0.2
python-dotenv==1.0.1
dkimpy==1.1.8
sqlmodel==0.0.22
pydantic[email]==2.9.2